+++ api-integration/src/App.js
@@ -1,57 +1,18 @@
-import { useReducer, useEffect } from 'react';
 import axios from 'axios';
+import useAsync from './hooks/useAsync';
 
-function reducer(state, action) {
-  switch (action.type) {
-    case 'LOADING':
-      return {
-        loading: true,
-        data: null,
-        error: null,
-      };
-    case 'SUCCESS':
-      return {
-        loading: false,
-        data: action.data,
-        error: null,
-      };
-    case 'ERROR':
-      return {
-        loading: false,
-        data: null,
-        error: action.error,
-      };
-    default:
-      return new Error(`Unhandled action type: ${action.type}`);
-  }
+async function getUsers() {
+  const res = await axios.get('https://jsonplaceholder.typicode.com/users');
+  return res.data;
 }
 
 function App() {
-  const [state, dispatch] = useReducer(reducer, {
-    loading: false,
-    data: null,
-    error: null,
-  });
-
-  const fetchUsers = async () => {
-    try {
-      dispatch({ type: 'LOADING' });
-
-      const res = await axios.get('https://jsonplaceholder.typicode.com/users');
-      dispatch({ type: 'SUCCESS', data: res.data });
-    } catch (e) {
-      dispatch({ type: 'ERROR' });
-    }
-  };
-
-  useEffect(() => {
-    fetchUsers();
-  }, []);
-
+  const [state, refetch] = useAsync(getUsers, [], true);
   const { data: users, loading, error } = state;
+
   if (loading) return <div>...loading 중</div>;
   if (error) return <div>에러가 발생했습니다.</div>;
-  if (!users) return null;
+  if (!users) return <button onClick={refetch}>불러오기</button>;
   return (
     <>
       <ul>
@@ -61,7 +22,7 @@ function App() {
           </li>
         ))}
       </ul>
-      <button onClick={fetchUsers}>다시 불러오기</button>
+      <button onClick={refetch}>다시 불러오기</button>
     </>
   );
 }
